+++ CHANGELOG.md
@@ -21,12 +21,14 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 * [`no-unused-state`]: avoid crashing on a class field function with destructured state ([#3568][] @ljharb)
 * [`no-unused-prop-types`]: allow using spread with object expression in jsx ([#3570][] @akulsr0)
 * Revert "[`destructuring-assignment`]: Handle destructuring of useContext in SFC" ([#3583][] [#2797][] @102)
+* [`prefer-read-only-props`]: add TS support ([#3593][] @HenryBrown0)
 
 ### Changed
 * [Docs] [`jsx-newline`], [`no-unsafe`], [`static-property-placement`]: Fix code syntax highlighting ([#3563][] @nbsp1221)
 * [readme] resore configuration URL ([#3582][] @gokaygurcan)
 * [Docs] [`jsx-no-bind`]: reword performance rationale ([#3581][] @gpoole)
 
+[#3593]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3593
 [#3583]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3583
 [#3582]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3582
 [#3581]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3581